Extract thread conversation file change helpers
