+++ src/components/content/ThreadConversation.vue
@@ -871,6 +871,20 @@
 import { computed, nextTick, onBeforeUnmount, onMounted, ref, watch } from 'vue'
 import type { UiFileChange, UiLiveOverlay, UiMessage, UiPlanStep, UiServerRequest } from '../../types/codex'
 import { useMobile } from '../../composables/useMobile'
+import {
+  aggregateFileChanges,
+  buildDiffViewerLines,
+  fileChangeDeltaParts,
+  fileChangeKey,
+  fileChangeOperationLabel,
+  fileChangeOperationTone,
+  fileChangeSummaryLabel,
+  fileChangeSummaryStatusParts,
+  formatFileChangeDelta,
+  formatFileChangeCountLabel,
+  type DiffViewerLine,
+  type TurnFileChangeSummary,
+} from './threadConversationFileChanges'
 
 import IconTablerArrowUp from '../icons/IconTablerArrowUp.vue'
 import IconTablerCopy from '../icons/IconTablerCopy.vue'
@@ -1159,10 +1173,6 @@ function isFileChangeSummaryExpanded(message: UiMessage): boolean {
   return expandedFileChangeSummaryIds.value.has(message.id)
 }
 
-function fileChangeKey(change: UiFileChange): string {
-  return `${change.path}\u0000${change.movedToPath ?? ''}`
-}
-
 function openDiffViewer(summary: TurnFileChangeSummary | null, change: UiFileChange): void {
   if (!summary) return
   activeDiffViewerSummary.value = summary
@@ -1401,20 +1411,6 @@ type McpElicitationField = {
   options: McpElicitationFieldOption[]
   defaultValue: string | number | boolean | string[]
 }
-type TurnFileChangeSummary = {
-  changes: UiFileChange[]
-  sourceMessageIds: string[]
-  source: 'assistant' | 'metadata'
-}
-type DiffViewerLineKind = 'meta' | 'hunk' | 'add' | 'remove' | 'context'
-type DiffViewerLine = {
-  key: string
-  kind: DiffViewerLineKind
-  oldLine: number | null
-  newLine: number | null
-  text: string
-}
-
 function isFilePath(value: string): boolean {
   if (!value || /[\r\n]/u.test(value)) return false
   if (value.endsWith('/') || value.endsWith('\\')) return false
@@ -1737,47 +1733,6 @@ function showForkResponseButton(message: UiMessage): boolean {
   return typeof forkableTurnIndexByAnchorId.value[message.id] === 'number'
 }
 
-function mergeFileChangeDiff(first: string, second: string): string {
-  if (!first) return second
-  if (!second || first === second) return first
-  return `${first}\n${second}`.trim()
-}
-
-function mergeFileChangeEntry(first: UiFileChange, second: UiFileChange): UiFileChange {
-  const operation = first.operation === 'add' || second.operation === 'add'
-    ? 'add'
-    : first.operation === 'delete' || second.operation === 'delete'
-      ? 'delete'
-      : 'update'
-  return {
-    path: second.path || first.path,
-    operation,
-    movedToPath: second.movedToPath ?? first.movedToPath ?? null,
-    diff: mergeFileChangeDiff(first.diff, second.diff),
-    addedLineCount: first.addedLineCount + second.addedLineCount,
-    removedLineCount: first.removedLineCount + second.removedLineCount,
-  }
-}
-
-function compareFileChanges(first: UiFileChange, second: UiFileChange): number {
-  const firstRank = first.operation === 'add' ? 0 : first.operation === 'update' ? 1 : 2
-  const secondRank = second.operation === 'add' ? 0 : second.operation === 'update' ? 1 : 2
-  if (firstRank !== secondRank) return firstRank - secondRank
-  const firstPath = `${first.path}\u0000${first.movedToPath ?? ''}`
-  const secondPath = `${second.path}\u0000${second.movedToPath ?? ''}`
-  return firstPath.localeCompare(secondPath)
-}
-
-function aggregateFileChanges(changes: UiFileChange[]): UiFileChange[] {
-  const byPath = new Map<string, UiFileChange>()
-  for (const change of changes) {
-    const key = `${change.path}\u0000${change.movedToPath ?? ''}`
-    const previous = byPath.get(key)
-    byPath.set(key, previous ? mergeFileChangeEntry(previous, change) : { ...change })
-  }
-  return Array.from(byPath.values()).sort(compareFileChanges)
-}
-
 const anchoredFileChangeSummaryByAnchorId = computed<Record<string, TurnFileChangeSummary>>(() => {
   const assistantAnchorIdByTurnKey = new Map<string, string>()
   const assistantSummaryByAnchorId = new Map<string, TurnFileChangeSummary>()
@@ -1878,96 +1833,6 @@ function readStandaloneFileChangeSummary(message: UiMessage): TurnFileChangeSumm
   return standaloneFileChangeSummaryByMessageId.value[message.id] ?? null
 }
 
-function fileChangeOperationLabel(change: UiFileChange): string {
-  if (change.operation === 'update' && change.movedToPath) {
-    return change.addedLineCount > 0 || change.removedLineCount > 0 ? 'Moved + edited' : 'Moved'
-  }
-  if (change.operation === 'add') return 'Added'
-  if (change.operation === 'delete') return 'Deleted'
-  return 'Edited'
-}
-
-function fileChangeOperationTone(change: UiFileChange): 'add' | 'delete' | 'update' | 'move' {
-  if (change.operation === 'update' && change.movedToPath) return 'move'
-  return change.operation
-}
-
-function formatFileChangeDelta(change: UiFileChange): string {
-  const parts: string[] = []
-  if (change.addedLineCount > 0) parts.push(`+${change.addedLineCount}`)
-  if (change.removedLineCount > 0) parts.push(`-${change.removedLineCount}`)
-  return parts.join(' ')
-}
-
-type FileChangeDeltaTone = 'add' | 'remove' | 'neutral'
-
-type FileChangeDeltaPart = {
-  tone: FileChangeDeltaTone
-  label: string
-}
-
-function buildFileChangeDeltaParts(addedCount: number, removedCount: number, fallbackLabel = ''): FileChangeDeltaPart[] {
-  const parts: FileChangeDeltaPart[] = []
-  if (addedCount > 0) parts.push({ tone: 'add', label: `+${addedCount}` })
-  if (removedCount > 0) parts.push({ tone: 'remove', label: `-${removedCount}` })
-  if (parts.length > 0) return parts
-  return fallbackLabel ? [{ tone: 'neutral', label: fallbackLabel }] : []
-}
-
-function fileChangeDeltaParts(change: UiFileChange): FileChangeDeltaPart[] {
-  return buildFileChangeDeltaParts(change.addedLineCount, change.removedLineCount)
-}
-
-function formatFileChangeCountLabel(count: number): string {
-  return count === 1 ? '1 file changed' : `${count} files changed`
-}
-
-function summarizeFileChangeKinds(summary: TurnFileChangeSummary | null): string {
-  if (!summary || summary.changes.length === 0) return ''
-  let added = 0
-  let deleted = 0
-  let edited = 0
-  let moved = 0
-
-  for (const change of summary.changes) {
-    if (change.operation === 'add') {
-      added += 1
-      continue
-    }
-    if (change.operation === 'delete') {
-      deleted += 1
-      continue
-    }
-    if (change.movedToPath) {
-      moved += 1
-      continue
-    }
-    edited += 1
-  }
-
-  const parts: string[] = []
-  if (edited > 0) parts.push(`${edited} edited`)
-  if (added > 0) parts.push(`${added} added`)
-  if (deleted > 0) parts.push(`${deleted} deleted`)
-  if (moved > 0) parts.push(`${moved} moved`)
-  return parts.join(', ')
-}
-
-function fileChangeSummaryLabel(summary: TurnFileChangeSummary | null): string {
-  if (!summary || summary.changes.length === 0) return 'Modified files'
-  const countLabel = formatFileChangeCountLabel(summary.changes.length)
-  const kindSummary = summarizeFileChangeKinds(summary)
-  return kindSummary ? `${countLabel} · ${kindSummary}` : countLabel
-}
-
-function fileChangeSummaryStatusParts(summary: TurnFileChangeSummary | null): FileChangeDeltaPart[] {
-  if (!summary || summary.changes.length === 0) return []
-  const totalAdded = summary.changes.reduce((sum, change) => sum + change.addedLineCount, 0)
-  const totalRemoved = summary.changes.reduce((sum, change) => sum + change.removedLineCount, 0)
-  const fallbackLabel = summary.changes.some((change) => change.movedToPath) ? 'Moved' : 'Ready'
-  return buildFileChangeDeltaParts(totalAdded, totalRemoved, fallbackLabel)
-}
-
 function displayFileChangePath(pathValue: string): string {
   const resolved = resolveRelativePath(pathValue, props.cwd)
   const normalizedCwd = normalizePathDots(normalizePathSeparators(props.cwd.trim()))
@@ -2001,108 +1866,6 @@ function inferDiffViewerLanguage(change: UiFileChange): string {
   const targetPath = change.movedToPath || change.path
   const extension = targetPath.split('.').pop()?.toLowerCase() ?? ''
   return CODE_LANGUAGE_ALIASES[extension] ?? extension ?? ''
-}
-
-function hasStructuredUnifiedDiff(change: UiFileChange): boolean {
-  return change.operation === 'update' && /^diff --git |^@@ |^--- |^\+\+\+ |^[ +-]|^\*\*\* (Move to:|End of File)/mu.test(change.diff)
-}
-
-function buildSyntheticDiffLines(change: UiFileChange): DiffViewerLine[] {
-  const normalized = change.diff.replace(/\r\n/g, '\n')
-  const lines = normalized.length > 0 ? normalized.split('\n') : []
-  if (lines.length > 0 && lines[lines.length - 1] === '') {
-    lines.pop()
-  }
-  return lines.map((line, index) => ({
-    key: `${fileChangeKey(change)}:synthetic:${index}`,
-    kind: change.operation === 'delete' ? 'remove' : 'add',
-    oldLine: change.operation === 'delete' ? index + 1 : null,
-    newLine: change.operation === 'delete' ? null : index + 1,
-    text: line,
-  }))
-}
-
-function buildUnifiedDiffLines(change: UiFileChange): DiffViewerLine[] {
-  const normalized = change.diff.replace(/\r\n/g, '\n')
-  const lines = normalized.length > 0 ? normalized.split('\n') : []
-  if (lines.length > 0 && lines[lines.length - 1] === '') {
-    lines.pop()
-  }
-
-  const output: DiffViewerLine[] = []
-  let oldLine = 0
-  let newLine = 0
-
-  for (const [index, line] of lines.entries()) {
-    const hunkMatch = line.match(/^@@\s+-(\d+)(?:,\d+)?\s+\+(\d+)(?:,\d+)?\s+@@/u)
-    if (hunkMatch) {
-      oldLine = Number(hunkMatch[1])
-      newLine = Number(hunkMatch[2])
-      output.push({
-        key: `${fileChangeKey(change)}:hunk:${index}`,
-        kind: 'hunk',
-        oldLine: null,
-        newLine: null,
-        text: line,
-      })
-      continue
-    }
-
-    if (line.startsWith('+') && !line.startsWith('+++')) {
-      output.push({
-        key: `${fileChangeKey(change)}:add:${index}`,
-        kind: 'add',
-        oldLine: null,
-        newLine,
-        text: line.slice(1),
-      })
-      newLine += 1
-      continue
-    }
-
-    if (line.startsWith('-') && !line.startsWith('---')) {
-      output.push({
-        key: `${fileChangeKey(change)}:remove:${index}`,
-        kind: 'remove',
-        oldLine,
-        newLine: null,
-        text: line.slice(1),
-      })
-      oldLine += 1
-      continue
-    }
-
-    if (line.startsWith(' ')) {
-      output.push({
-        key: `${fileChangeKey(change)}:context:${index}`,
-        kind: 'context',
-        oldLine,
-        newLine,
-        text: line.slice(1),
-      })
-      oldLine += 1
-      newLine += 1
-      continue
-    }
-
-    output.push({
-      key: `${fileChangeKey(change)}:meta:${index}`,
-      kind: 'meta',
-      oldLine: null,
-      newLine: null,
-      text: line,
-    })
-  }
-
-  return output
-}
-
-function buildDiffViewerLines(change: UiFileChange | null): DiffViewerLine[] {
-  if (!change || !change.diff.trim()) return []
-  if (hasStructuredUnifiedDiff(change)) {
-    return buildUnifiedDiffLines(change)
-  }
-  return buildSyntheticDiffLines(change)
 }
 
 const activeDiffViewerLines = computed<DiffViewerLine[]>(() => buildDiffViewerLines(activeDiffViewerChange.value))
